+++ allauth/socialaccount/providers/battlenet/urls.py
@@ -1,0 +1,4 @@
+from allauth.socialaccount.providers.oauth2.urls import default_urlpatterns
+from .provider import BattleNetProvider
+
+urlpatterns = default_urlpatterns(BattleNetProvider)