+++ test_settings.py
@@ -71,6 +71,7 @@ INSTALLED_APPS = (
     'allauth.socialaccount.providers.asana',
     'allauth.socialaccount.providers.baidu',
     'allauth.socialaccount.providers.basecamp',
+    'allauth.socialaccount.providers.battlenet',
     'allauth.socialaccount.providers.bitbucket',
     'allauth.socialaccount.providers.bitly',
     'allauth.socialaccount.providers.coinbase',